@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.3
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -30,18 +30,19 @@ dependencies:
   flutter:
     sdk: flutter
   adaptive_theme: ^3.1.1
-  flutterfire_ui: ^0.4.3+20
+  firebase_ui_auth: ^1.4.2
   firebase_core: ^2.7.1
-  google_sign_in: ^5.2.1
-  flutter_svg: ^1.0.0
+  google_sign_in: ^6.1.4
+  flutter_svg: ^2.0.0+1
   firebase_auth: ^4.2.10
   file_selector: ^0.9.2
   image_picker: ^0.8.5+3
   url_launcher: ^6.1.5
-  intl: ^0.17.0
-  flutter_extensions: ^4.2.0
+  flutter_extensions: ^5.0.0
   visual_editor:
-    git: https://github.com/visual-space/visual-editor.git
+    git:
+      url: https://github.com/visual-space/visual-editor.git
+      ref: develop
   tuple: ^2.0.0
   cupertino_icons: ^1.0.5
   cloud_firestore: ^4.4.4
@@ -55,20 +56,28 @@ dependencies:
   crop_your_image: ^0.7.2
   uuid: ^3.0.6
   algolia: ^1.1.1
-  math_keyboard: ^0.1.7
+  math_keyboard: ^0.2.0
   flutter_native_splash: ^2.2.19
-  flutter_math_fork: ^0.6.3+1
+  flutter_math_fork: ^0.7.1
   firebase_analytics: ^10.1.5
   path: ^1.8.2
-  video_player_macos: 1.0.6
+  video_player_macos: 2.0.1
   scrollable_positioned_list: ^0.2.3
   azlistview: ^2.0.0
+  collection: ^1.17.1
+  firebase_ui_oauth_google: ^1.2.2
+  firebase_ui_oauth_apple: ^1.2.2
 
 dependency_overrides:
+  intl: ^0.18.0
   flutter_colorpicker:
     git:
       url: https://github.com/mchome/flutter_colorpicker
       ref: master
+  macos_ui:
+    git:
+      url: https://github.com/macosui/macos_ui
+      ref: dev
 
 dev_dependencies:
   flutter_test:
