@@ -24,7 +24,7 @@ class SliderLabel extends StatelessWidget {
     }
     final modifier = appearance.infoModifier(value);
     widgets.add(
-      Text('$modifier', style: appearance.infoMainLabelStyle),
+      Text(modifier, style: appearance.infoMainLabelStyle),
     );
     if (appearance.infoBottomLabelText != null) {
       widgets.add(Text(
